Fold the Dynamic wire constants into the codec
DynamicWire held three members that only DynamicColumnCodec used: the
flattened serialization version, the type-count ceiling, and the
discriminator-width rule. A separate file for them added a hop without
adding a boundary, so move them into the codec — the version and the
ceiling as private constants, the width rule as an internal static
method the tests still reach.

The width test moves with it, from DynamicTypeInferenceTests to
DynamicColumnCodecTests, where it now belongs.
